Fixed unkown character error and working on reorganizing the front end code with the developers code
