--- constants/colors.ts
@@ -1,8 +1,0 @@
-export const colors = {
-  inActiveColor: "#8d8b8b",
-  light: "#fff",
-  tabBarBackgroundColor: "#101010",
-  pageHeaderBackgroundColor: "#101010",
-  successColor: "#00E188",
-  warning: "#FF9600",
-} as const;
