--- constants/dimension.ts
@@ -1,3 +1,0 @@
-export const dimension = {
-  headerHeight: 80,
-} as const;
