+++ src/constants/dimension.ts
@@ -1,0 +1,3 @@
+export const dimension = {
+  headerHeight: 80,
+} as const;
