+++ src/icons/LockIcon.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import Svg, { SvgProps, Path } from "react-native-svg";
-import { colors } from "../../constants/colors";
+import { colors } from "../constants/colors";
 
 function LockICon(props: SvgProps) {
   return (
