--- test/readme.txt
@@ -1,1 +1,0 @@
-Mocha Tests
